+++ apyfal/storage/aws.py
@@ -28,7 +28,8 @@ class AWSStorage(_BucketStorage):
 
     Args:
         storage_type (str): Cloud service provider name. Default to "AWS".
-        config (str or apyfal.configuration.Configuration): Configuration file path or instance.
+        config (str or apyfal.configuration.Configuration or file-like object):
+            Can be Configuration instance, apyfal.storage URL, paths, file-like object.
             If not set, will search it in current working directory, in current
             user "home" folder. If none found, will use default configuration values.
         client_id (str): AWS Access Key ID.