+++ apyfal/storage/http.py
@@ -15,7 +15,8 @@ class HTTPStorage(_Storage):
 
     Args:
         storage_type (str): Type of storage. Default to "HTTP".
-        config (str or apyfal.configuration.Configuration): Configuration file path or instance.
+        config (str or apyfal.configuration.Configuration or file-like object):
+            Can be Configuration instance, apyfal.storage URL, paths, file-like object.
             If not set, will search it in current working directory, in current
             user "home" folder. If none found, will use default configuration values.
     """