+++ tests/test_host_csp.py
@@ -642,7 +642,6 @@ def run_full_real_test_sequence(host_type, environment,
         support_stop_restart (bool): If True support pause instance
             and restart.
     """
-    # TODO: Update to use 'host.host_type' in config
     from apyfal.configuration import Configuration
 
     # Skip if no correct configuration with this host_type
@@ -724,5 +723,3 @@ def run_full_real_test_sequence(host_type, environment,
     finally:
         if not use_full_images:
             apyfal._utilities.check_url = utilities_check_url
-
-        # TODO: CSP clean up...